chore: relocate tools and agent
